003 Code/requirements.txt
keyboard==0.13.5
pyserial==3.5
pyzmq==25.0.0
tornado==6.2
wlkata-mirobot-python==0.1.14
aiohttp==3.8.3
aioice==0.7.6
aiortc==1.4.0
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
av==10.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
dnspython==2.3.0
Flask==2.2.2
frozenlist==1.3.3
google-crc32c==1.5.0
idna==3.4
inputs==0.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
netifaces==0.11.0
numpy==1.24.2
opencv-python==4.7.0.68
pybase64==1.2.3
pycparser==2.21
pyee==9.0.4
pylibsrtp==0.8.0
pynmea2==1.19.0
pyOpenSSL==23.0.0
pyserial==3.5
pyserial-asyncio==0.6
requests==2.28.2
rtcbot==0.2.4
SoundCard==0.4.2
typing_extensions==4.4.0
urllib3==1.26.14
websockets==10.4
Werkzeug==2.2.2
yarl==1.8.2